Backbone for Wordle Creation landing Page
This page will allow users to create 'Wordle games' for other users to try and play. The user will have the ability to get given a random word, or choose a word (not implemented, will be making sure the word they choose is in the database). Currently, there is no database insertions created for this.